+++ src/Touch.cpp
@@ -208,11 +208,11 @@ void Touch::startDragCursors(int x, int y)
 {
 	state.setCursorDrag(true);
 	vector<int> coordenates = state.getGridCoordenates();
-	vector<double> cursor;
-	cursor.push_back((double)(x - coordenates[0] - 5 * state.getPixelsPerDivision()) / state.getPixelsPerDivision());
-	cursor.push_back((double)(x - coordenates[0] - 5 * state.getPixelsPerDivision()) / state.getPixelsPerDivision());
-	cursor.push_back((double)(coordenates[2] + 4 * state.getPixelsPerDivision() - y) / state.getPixelsPerDivision());
-	cursor.push_back((double)(coordenates[2] + 4 * state.getPixelsPerDivision() - y) / state.getPixelsPerDivision());
+	vector<float> cursor;
+	cursor.push_back((float)(x - coordenates[0] - 5 * state.getPixelsPerDivision()) / state.getPixelsPerDivision());
+	cursor.push_back((float)(x - coordenates[0] - 5 * state.getPixelsPerDivision()) / state.getPixelsPerDivision());
+	cursor.push_back((float)(coordenates[2] + 4 * state.getPixelsPerDivision() - y) / state.getPixelsPerDivision());
+	cursor.push_back((float)(coordenates[2] + 4 * state.getPixelsPerDivision() - y) / state.getPixelsPerDivision());
 	configuration.setCursor(cursor);
 }
 
@@ -255,9 +255,9 @@ void Touch::resetDrag(void)
 void Touch::dragCursor(int x, int y)
 {
 	vector<int> coordenates = state.getGridCoordenates();
-	vector<double> cursor = configuration.getCursor();
-	cursor[1] = (double)(x - coordenates[0] - 5 * state.getPixelsPerDivision()) / state.getPixelsPerDivision();
-	cursor[3] = (double)(coordenates[2] + 4 * state.getPixelsPerDivision() - y) / state.getPixelsPerDivision();
+	vector<float> cursor = configuration.getCursor();
+	cursor[1] = (float)(x - coordenates[0] - 5 * state.getPixelsPerDivision()) / state.getPixelsPerDivision();
+	cursor[3] = (float)(coordenates[2] + 4 * state.getPixelsPerDivision() - y) / state.getPixelsPerDivision();
 	configuration.setCursor(cursor);
 }
 
@@ -269,7 +269,7 @@ void Touch::dragOffset(int x, int y)
 		configuration.setOffset(state.getSelectedChannel(), configuration.getOffset(state.getSelectedChannel()) + configuration.getVerticalScaleValue(state.getSelectedChannel()) * (initialOffset[1] - y) / state.getPixelsPerDivision());
 		initialOffset[1] = y;
 	}
-	configuration.setDelay(configuration.getDelay() + (double)(x - initialOffset[0]) / state.getPixelsPerDivision());
+	configuration.setDelay(configuration.getDelay() + (float)(x - initialOffset[0]) / state.getPixelsPerDivision());
 	initialOffset[0] = x;
 	state.setInitialOffset(initialOffset);
 }
@@ -289,9 +289,9 @@ void Touch::dragHorizontalScale(int x)
 	int levels = (x - state.getInitialHorizontalScale()) / state.getPixelsPerDivision();
 	if(levels != 0)
 	{
-		double oldScale = configuration.getHorizontalScaleValue();
+		float oldScale = configuration.getHorizontalScaleValue();
 		configuration.setHorizontalScale((Configuration::HorizontalScales)(configuration.getHorizontalScale() + levels));
-		double newScale = configuration.getHorizontalScaleValue();
+		float newScale = configuration.getHorizontalScaleValue();
 		if(configuration.getMode() == Configuration::STOP)
 		{
 			configuration.setStep(configuration.getStep() * newScale / oldScale);
@@ -303,8 +303,8 @@ void Touch::dragHorizontalScale(int x)
 void Touch::dragTriggerLevel(int x, int y)
 {
 	vector<int> coordenates = state.getGridCoordenates();
-	configuration.setTriggerHoldOff(configuration.getHorizontalScaleValue() * ((double)(x - coordenates[0])) / state.getPixelsPerDivision());
-	configuration.setTriggerLevel(((double) (coordenates[2] + 4 * state.getPixelsPerDivision() - y)) / state.getPixelsPerDivision() - configuration.getOffset(configuration.getTriggerChannel()) / configuration.getVerticalScaleValue(configuration.getTriggerChannel()));
+	configuration.setTriggerHoldOff(configuration.getHorizontalScaleValue() * ((float)(x - coordenates[0])) / state.getPixelsPerDivision());
+	configuration.setTriggerLevel(((float) (coordenates[2] + 4 * state.getPixelsPerDivision() - y)) / state.getPixelsPerDivision() - configuration.getOffset(configuration.getTriggerChannel()) / configuration.getVerticalScaleValue(configuration.getTriggerChannel()));
 }
 
 void Touch::resetButtons(void)
